Hoist Toolchain up into top-level toolchain module
